+++ src/test/kotlin/racingcar/model/round/RoundTest.kt
@@ -1,10 +1,9 @@
-package racingcar.model
+package racingcar.model.round
 
 import org.junit.jupiter.api.Assertions.assertDoesNotThrow
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
-import racingcar.model.round.Round
 import java.lang.IllegalArgumentException
 
 internal class RoundTest {
